devops: apply patch failsafe
When running prepare_checkout.sh on a newer BASE_REVISION "git apply"
fails if there are conflicts. This use case is useful when rebasing
bootstrap.patch to a new commit.

Now when there is an environment variable APPLY_PATCH_FAILSAFE which
when defined makes "git apply" not to fail if there are conflicts.
In addition, conflicts are stored in a file called 'conflicts' in CHECKOUT_PATH.

The patch also allows to pass a second argument to
./prepare_checkout.sh, which will overwrite the commit id value of
BASE_REVISION in UPSTREAM_CONFIG.sh.

This feature is only support for WebKit browser.